refactor: separate helper functions
