+++ apps/dashboard/features/governance/utils/index.ts
@@ -1,0 +1,5 @@
+export * from "@/features/governance/utils/calculateDaysDiff";
+export * from "@/features/governance/utils/formatVotes";
+export * from "@/features/governance/utils/getTimeText";
+export * from "@/features/governance/utils/getProposalStatus";
+export * from "@/features/governance/utils/getProposalState";
