+++ apps/dashboard/shared/hooks/index.ts
@@ -2,7 +2,6 @@ export * from "@/shared/hooks/useActiveSupply";
 export * from "@/shared/hooks/useAverageTurnout";
 export * from "@/shared/hooks/useDaoData";
 export * from "@/shared/hooks/useDelegatedSupply";
-export * from "@/shared/hooks/useProposals";
 export * from "@/shared/hooks/useTimeSeriesData";
 export * from "@/shared/hooks/useVotes";
 export * from "@/shared/hooks/useScreenSize";
